% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_GGMtest.R
\name{adj_GGMtest}
\alias{adj_GGMtest}
\title{Create adjacency matrix for GGMtest object}
\usage{
adj_GGMtest(object)
}
\arguments{
\item{object}{GGMtest object}
}
\value{
An adjacency matrix with TRUE for all rejected edges. All other edges are set to FALSE.
}
\description{
Create adjacency matrix for GGMtest object
}
\seealso{
\code{\link{GGMtest}}
}
