% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_intervals.R
\name{confint.GGMtest}
\alias{confint.GGMtest}
\title{Create the confidence intervals for a GGMtest object}
\usage{
\method{confint}{GGMtest}(object)
}
\description{
Create the confidence intervals for a GGMtest object
}
