% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMtest.R
\name{GGMtest}
\alias{GGMtest}
\title{Estimates conditional independence hypothsis for a Gaussian Graphical Model}
\usage{
GGMtest(
  data = X,
  edges = S,
  beta_0 = rep(0, dim(S)[1]),
  alpha = 0.05,
  nbootstrap = 100,
  nuisance_estimaton = "lasso",
  s = 1,
  exponent = 1,
  penalty = list(c = 1.1),
  k_fold = 1,
  rnd_seed = NULL
)
}
\description{
Estimates conditional independence hypothsis for a Gaussian Graphical Model
}
