% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMtest2.R
\name{GGMtest2}
\alias{GGMtest2}
\title{GGMtest2}
\usage{
GGMtest2(
  data = X,
  edges = S,
  nuisance_estimaton = "lasso",
  method = "robust",
  DML_method = "DML2",
  k_fold = 1,
  penalty = list(c = 1.1)
)
}
\arguments{
\item{data}{Dataset: either matrix or dataframe}

\item{edges}{Matrix of edges for testing: each row specifies an edge}

\item{nuisance_estimaton}{Method for nuisance parameter estimation from 'lasso', 'post-lasso' or 'sqrt-lasso'}

\item{method}{Method for point estimation, either 'root' or 'partialling out'}

\item{DML_method}{Method for point estimation, either 'DML2' or 'DML1'}

\item{k_fold}{Parameter for K-fold estimation. Default is k_fold = 1.}

\item{penalty}{Additional coefficient for the penalty term. Default value is c = 1.1.}
}
\value{
A list with components
\item{estimates}{A vector of point estimates.}
\item{edge_list}{The matrix containing the corresponding edges (equal to input).}
\item{sigma_est}{Estimates of the standard deviation.}
\item{psi_est}{Estimates of the score vector.}
\item{additional_parameters}{Additional parameters.}
}
\description{
Testing conditional independence hypothesis for a gaussian graphical model.
}
\examples{
library("huge")
library("igraph")
library("GGMtest")

set.seed(42)

# generate data (different graph structures: "random", "hub", "cluster", "band" and "scale-free")
L <- huge.generator(n = 100, d = 10, graph = "cluster", g = 4)

# true Graph
true_graph <- graph_from_adjacency_matrix(as.matrix(L$theta), mode='undirected', diag=FALSE)
plot(true_graph, usearrows = FALSE, label=1:10, displaylabels=TRUE, main = "True Graph",layout= layout.fruchterman.reingold, edge.width = 2, edge.color = "black")

# index pairs for inference
S <- matrix(c(1,2,2,3,4,5), byrow = TRUE, ncol = 2)

# perform test
ggm_model <- GGMtest2(data = L$data,edges = S,nuisance_estimaton = "lasso")

# Create Confidence Region:
create_CR(ggm_model)


}
\seealso{
\code{\link{confint.GGMtest}} for confidence intervals, \code{\link{plot_GGMtest}} for plotting options
 and \code{\link{adj_GGMtest}} for the adjacency matrix
}
