% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GGMtest.R
\name{plot.GGMtest}
\alias{plot.GGMtest}
\title{Plot confidence intervals for GGMtest object}
\usage{
\method{plot}{GGMtest}(object, edges = object$edge_list[1, ])
}
\description{
Plot confidence intervals for GGMtest object
}
