% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GGMtest.R
\name{plot_GGMtest}
\alias{plot_GGMtest}
\title{Plot confidence intervals for GGMtest objects}
\usage{
plot_GGMtest(object, edges = object$edge_list[1, ], ...)
}
\arguments{
\item{object}{GGMtest Object}

\item{edges}{List of edges to plot. Ignores other edges which were not tested in the object.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\description{
Only usable for s=1 or s=2 and exponent=2.
}
\seealso{
\code{\link{GGMtest}}
}
