% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMtest2.R
\name{create_CR}
\alias{create_CR}
\title{Create Confidence regions for a `GGMtest2` object.}
\usage{
create_CR(model, alpha = 0.05, B = 500, s = 1, exp = 1)
}
\arguments{
\item{model}{The `GGMtest2` object.}

\item{alpha}{The corresponding level.}

\item{B}{The number of used Bootstrap repetitions.}

\item{s}{The number of s-sparse combinations.}

\item{exp}{The corresponding exponent of the s-sparse sets.}
}
\value{
A list with the following components.
}
\description{
Create Confidence regions for a `GGMtest2` object.
}
