% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGMtest2.R
\name{create_CR}
\alias{create_CR}
\title{Create Confidence regions for a `GGMtest2` object.}
\usage{
create_CR(model, null_hyp = 0, alpha = 0.05, B = 500, s = 1, exp = 1)
}
\arguments{
\item{model}{The `GGMtest2` object.}

\item{null_hyp}{A vector of null hypothesis values for the coefficients: Default is 0 for conditional independence}

\item{alpha}{The corresponding level.}

\item{B}{The number of used Bootstrap repetitions.}

\item{s}{The number of s-sparse combinations.}

\item{exp}{The corresponding exponent of the s-sparse sets.}
}
\value{
A list with the following components.
\item{estimates}{A vector of point estimates.}
\item{edge_list}{The matrix containing the corresponding edges (equal to input).}
\item{pvalue_max}{P-Value of the maximum statistic.}
\item{pvalue_sphere}{P-Value of the approx. Sphere.}
\item{hyp_max}{`FALSE` if the hypthesis is rejected with the maximum statistic.}
\item{hyp_sphere}{`FALSE` if the hypthesis is rejected with the approx. Sphere.}
\item{vol_max}{Volume of the maximum statistic.}
\item{vol_sphere}{Volume of  the approx. Sphere.}
\item{quantiles}{Estimates quantiles.}
}
\description{
Create Confidence regions for a `GGMtest2` object.
}
